fix calculating vertical image while cropping
update itextpdf library
add retrofit write timeout

fix bugs:
https://fabric.io/rhino-family/android/apps/com.mobile.otrcapitalllc/issues/5ab253e48cb3c2fa63932781
https://fabric.io/rhino-family/android/apps/com.mobile.otrcapitalllc/issues/5abf981136c7b235272e8567
https://fabric.io/rhino-family/android/apps/com.mobile.otrcapitalllc/issues/5ac4f90136c7b23527856129

+++ app/src/main/java/com/mobile/otrcapitalllc/Activities/CropImage.java
@@ -233,7 +233,7 @@ public class CropImage extends BaseActivity {
         int scaledHeight = editImageView.getMeasuredHeight();
 
 
-        boolean isVertical = intrinsicWidth / intrinsicHeight <= scaledWidth / scaledHeight;
+        boolean isVertical = ((double)intrinsicWidth) / ((double)intrinsicHeight) <= ((double)scaledWidth) / ((double)scaledHeight);
         //Find the ratio of the original image to the scaled image
         double heightRatio = (double) intrinsicHeight / (double) scaledHeight;
         double widthRatio = (double) intrinsicWidth / (double) scaledWidth;
